Moved some cryptographic operations in the shared codebase
