+++ src/oracle/proxy/Dockerfile
@@ -6,6 +6,6 @@ COPY ./proxy/package.json /proxy/
 
 RUN npm install
 
-COPY ./proxy/index.js ./proxy/encode.js ./proxy/decode.js ./shared/logger.js /proxy/
+COPY ./proxy/index.js ./proxy/encode.js ./proxy/decode.js ./shared/logger.js ./shared/crypto.js /proxy/
 
 ENTRYPOINT ["node", "index.js"]
