+++ src/oracle/proxy/decode.js
@@ -224,6 +224,5 @@ module.exports = function (isKeygen, round, value) {
   const tokenizer = Tokenizer(value)
   const roundNumber = parseInt(round[round.length - 1])
   const decoder = (isKeygen ? keygenDecoders : signDecoders)[roundNumber]
-  const decoded = JSON.stringify(decoder(tokenizer))
-  return decoded
+  return JSON.stringify(decoder(tokenizer))
 }
