+++ src/oracle/tss-keygen/keygen.js
@@ -1,10 +1,9 @@
 const exec = require('child_process')
 const fs = require('fs')
-const crypto = require('crypto')
-const bech32 = require('bech32')
 
 const logger = require('./logger')
 const { connectRabbit, assertQueue } = require('./amqp')
+const { publicKeyToAddress } = require('./crypto')
 
 const { RABBITMQ_URL, PROXY_URL } = process.env
 
@@ -63,18 +62,4 @@ main()
 
 async function confirmKeygen (keysFile) {
   exec.execSync(`curl -X POST -H "Content-Type: application/json" -d @"${keysFile}" "${PROXY_URL}/confirmKeygen"`, { stdio: 'pipe' })
-}
-
-function publicKeyToAddress ({ x, y }) {
-  const compact = (parseInt(y[y.length - 1], 16) % 2 ? '03' : '02') + padZeros(x, 64)
-  const sha256Hash = crypto.createHash('sha256').update(Buffer.from(compact, 'hex')).digest('hex')
-  const hash = crypto.createHash('ripemd160').update(Buffer.from(sha256Hash, 'hex')).digest('hex')
-  const words = bech32.toWords(Buffer.from(hash, 'hex'))
-  return bech32.encode('tbnb', words)
-}
-
-function padZeros (s, len) {
-  while (s.length < len)
-    s = '0' + s
-  return s
 }
