+++ src/oracle/tss-sign/Dockerfile
@@ -10,7 +10,7 @@ COPY ./tss-sign/package.json /tss/
 
 RUN npm install --no-optional
 
-COPY ./tss-sign/sign-entrypoint.sh ./tss-sign/signer.js ./tss-sign/tx.js ./shared/logger.js ./shared/amqp.js /tss/
+COPY ./tss-sign/sign-entrypoint.sh ./tss-sign/signer.js ./tss-sign/tx.js ./shared/logger.js ./shared/amqp.js ./shared/crypto.js /tss/
 
 COPY --from=tss /tss/target/release/gg18_sign_client /tss/
